fix int overflow

+++ src/turbomind/models/llama/llama_kernels.h
@@ -86,7 +86,7 @@ public:
         int  stride_seq;
         int  stride_head;
         bool use_seqlens       = false;
-        int  batch_seqs_offset = 0;
+        size_t  batch_seqs_offset = 0;
         T**  batch_seqs        = nullptr;
     };
